JBR-339 refactor money

+++ src/main/java/com/jbr/middletier/money/dto/ArchiveOrReportRequestDTO.java
@@ -1,23 +1,23 @@
 package com.jbr.middletier.money.dto;
 
 public class ArchiveOrReportRequestDTO {
-    private long year;
-    private long month;
+    private int year;
+    private int month;
 
     public ArchiveOrReportRequestDTO() {
 
     }
 
-    public ArchiveOrReportRequestDTO(long year, long month) {
+    public ArchiveOrReportRequestDTO(int year, int month) {
         this.year = year;
         this.month = month;
     }
 
-    public long getYear() { return this.year; }
+    public int getYear() { return this.year; }
 
-    public void setYear(long year) { this.year = year; }
+    public void setYear(int year) { this.year = year; }
 
-    public long getMonth() { return this.month; }
+    public int getMonth() { return this.month; }
 
-    public void setMonth(long month) { this.month = month; }
+    public void setMonth(int month) { this.month = month; }
 }
